refactor redeem_voucher
